// src/app/core/interfaces/project.interface.ts
export interface ProjectDTO {
  id: number;
  name: string;
  areaId: number;
  areaName: string;
  cancellation: string;
  cityId: string;
  cityName: string;
  quota: number;
  responsible: string;
  status: string;
  typeId: string;
  typeName: string;

  /*
  idOrganization: id: number;
  agreementDate: string;
  status: string; //"2024-04-12\"
  agreementNumber: string; //"CCUQROO-01-2024"
  */


  /*
  \"areaId\" : \"527\",\n    
  \"name\" : \"ANALISIS Y MONITOREO DE MEDIOS DIRECCION DE COMUNICACION SOCIAL\",\n    
  \"responsible\" : \"LIC. IVAN OCAMPO RODRIGUEZ\",\n    
  \"typeId\" : \"01\",\n    
  \"quota\" : 10,\n    
  \"cityId\" : \"218\",\n    
  \"status\" : \"ACT\",\n    
  \"cancellation\" : null\n}",

  \"idOrganization\" : 306,\n    
  \"agreementDate\" : \"2024-04-12\",\n    
  \"status\" : \"VIG\",\n    
  \"agreementNumber\" : \"CCUQROO-01-2024\"\n}",
  */

}